requirements.txt
PyJWT
aiofiles
chromadb>=0.4.13
emoji
fastapi
httpx
httpx_sse
ijson
importlib_metadata>=4.6; python_version<'3.10'
lancedb>=0.2
packaging
panel==1.4.2
pyarrow
pydantic-core
pydantic-settings>=2
pydantic>=2
pymupdf>=1.23.6
python-docx
python-multipart
python-pptx
questionary
redis
rich
sqlalchemy>=2
starlette
tiktoken
tomlkit
typer
uvicorn
